Revert "fix create negative address"
This reverts commit a373ab3c49cf87792278b427c2845ec5b25d2ec7.

+++ demo/src/main/java/global/eska/ddk/keygen/demo/DemoCommandLineRunner.java
@@ -9,7 +9,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 @Slf4j
 @Component
@@ -45,7 +45,7 @@ public class DemoCommandLineRunner implements CommandLineRunner {
         log.info("secret key: {}", keyPair.getSecretKey());
         log.info("secret key hex: {}", keyPair.getSecretKeyHex());
 
-        BigInteger address = accountCreator.getAddressByPublicKey(keyPair.getPublicKey());
+        BigDecimal address = accountCreator.getAddressByPublicKey(keyPair.getPublicKey());
 
         log.info("DDK address generated successful: {}", address);
     }
